+++ R/topography-files.R
@@ -57,17 +57,13 @@ gebco08_files <- function(all = FALSE, ...) {
 #' @rdname topography-files
 ramp_files <- function(all = FALSE, ...) {
   pattern <- if (all) {
-    # ASCII grids contain fields for latitude, longitude, elevation relative to
-    # the WGS84 ellipsoid, and elevation relative to the OSU91A geoid. Data are
-    # represented in decimal degrees, from -180 degrees (west) to 180 degrees
-    # (east) longitude; and -90 to -60 (south) degrees latitude.
-    #
-    # ARC/INFO coverages of RAMP DEMs are organized into individual directories
-    # for each resolution (1 km, 400 m, and 200 m) and geoid/ellipsoid.
-    "sidads.colorado.edu/pub/DATASETS/nsidc0082_radarsat_dem_v02/200M/BINARY"
+    "sidads.colorado.edu/pub/DATASETS/nsidc0082_radarsat_dem_v02/200M/ARCINFO"
+
     } else {
-      "sidads.colorado.edu/pub/DATASETS/nsidc0082_radarsat_dem_v02/200M/BINARY.*bin$"
+      "sidads.colorado.edu/pub/DATASETS/nsidc0082_radarsat_dem_v02/200M/ARCINFO/osu91a200m/.*w001001.adf$"
+
     }
+#  print(pattern)
   topo_files_generic(pattern)
 }